moved some common functions up
