+++ src/common.ts
@@ -372,6 +372,30 @@ export function parseHexOrDecInt(str: string): number {
     return str.startsWith('0x') ? parseInt(str.substring(2), 16) : parseInt(str, 10);
 }
 
+export function parseHostPort(hostPort: string) {
+    let port: number;
+    let host = '127.0.0.1';
+    const match = hostPort.match(/(.*)\:([0-9]+)/);
+    if (match) {
+        host = match[1] ? match[1] : host;
+        port = parseInt(match[2], 10);
+    } else {
+        if (hostPort.startsWith(':')) {
+            hostPort = hostPort.slice(1);
+        }
+        port = parseInt(hostPort, 10);
+    }
+    return { port: port, host: host };
+}
+
+export function getNonce() {
+    let text = '';
+    const possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+    for (let i = 0; i < 32; i++) {
+        text += possible.charAt(Math.floor(Math.random() * possible.length));
+    }
+    return text;
+}
 export class ResettableInterval {
     protected intervalId: NodeJS.Timeout;
     protected args: any[];
